the spotted anomaly of the shapeless <!:< operator
This is an abstract syntax mechanisation of Launchbury's famous
operational semantics for lazy evaluation. The file
src/launchbury/Test.scala contains an object "tester" with an example
where <:!< fails to work correctly. bll in line 8 has no BaseVal[_] in
its ancestors. Yet, shapeless throws an unexpected which relates to the
implicit parameter of BaseLetLam (line 195 in src/lops/BaseExp.scala)
that is to enforce inconsitency between BaseLetLam and BaseVal.